refactor(aionrs): isolate process containment
- Move platform-specific command containment out of CommandRunner
- Keep runner code focused on process execution and output capture
- Leave Windows Job Object FFI in its dedicated module
